refactor gui.go
